backend/requirements.txt
async-timeout 5.0.1
blinker       1.9.0
cffi          1.17.1
click         8.2.1
colorama      0.4.6
cryptography  45.0.5
Flask         3.1.1
flask-cors    6.0.1
itsdangerous  2.2.0
Jinja2        3.1.6
MarkupSafe    3.0.2
numpy         2.2.6
opencv-python 4.11.0.86
pillow        11.3.0
pip           25.1.1
psutil        7.0.0
pycparser     2.22
PyJWT         2.10.1
python-magic  0.4.27
pywin32       310
redis         6.2.0
setuptools    58.1.0
tqdm          4.67.1
Werkzeug      3.1.3
gunicorn